Refactor dev and prod server configuration
